refactor: replace old Path with new Path domain
